♻️ Refactor test framework state classes to reduce duplication Extracted common test execution logic from ESpec and ExUnit State classes into a shared TestCommandLineState base class. - Add HasCommandLine interface for configurations that provide commandLine() - Create generic TestCommandLineStatebase class for test frameworks - Update ESpec and ExUnit configurations to implement HasCommandLine - Simplify ESpec and ExUnit State classes to only specify framework name - Add WSL-compatible process creation using runWithModalProgressBlocking
This reduces code duplication between test framework implementations while maintaining type safety and avoiding impact on configurations that don't need commandLine() (e.g., erl.Configuration).
